+++ src/arithmetic/algebraic_expression/algebraic_expression.c
@@ -14,51 +14,9 @@
 
 #include <assert.h>
 
-static AlgebraicExpression *_AlgebraicExpression_OperationRemoveRightmostChild
-(
-	AlgebraicExpression *root  // Root from which to remove a child.
-) {
-	assert(root);
-	if(root->type != AL_OPERATION) return NULL;
-
-	// No child nodes to remove.
-	if(AlgebraicExpression_ChildCount(root) == 0) return NULL;
-
-	// Remove rightmost child.
-	AlgebraicExpression *child = array_pop(root->operation.children);
-	return child;
-}
-
-static AlgebraicExpression *_AlgebraicExpression_OperationRemoveLeftmostChild
-(
-	AlgebraicExpression *root   // Root from which to remove a child.
-) {
-	assert(root);
-	if(root->type != AL_OPERATION) return NULL;
-
-	uint child_count = AlgebraicExpression_ChildCount(root);
-	// No child nodes to remove.
-	if(child_count == 0) return NULL;
-
-	// Remove leftmost child.
-	AlgebraicExpression *child = root->operation.children[0];
-
-	// Shift left by 1.
-	for(uint i = 0; i < child_count - 1; i++) {
-		root->operation.children[i] = root->operation.children[i + 1];
-	}
-	array_pop(root->operation.children);
-
-	return child;
-}
-
-static bool _AlgebraicExpression_IsMultiplicationNode(const AlgebraicExpression *node) {
-	return (node->type == AL_OPERATION && node->operation.op == AL_EXP_MUL);
-}
-
-static bool _AlgebraicExpression_IsAdditionNode(const AlgebraicExpression *node) {
-	return (node->type == AL_OPERATION && node->operation.op == AL_EXP_ADD);
-}
+//------------------------------------------------------------------------------
+// Static internal functions.
+//------------------------------------------------------------------------------
 
 // Locate the left most node in `exp`
 static AlgebraicExpression *_leftMostNode(AlgebraicExpression *exp) {
@@ -76,212 +34,6 @@ static AlgebraicExpression *_rightMostNode(AlgebraicExpression *exp) {
 		right_most = LAST_CHILD(right_most);
 	}
 	return right_most;
-}
-
-static bool __AlgebraicExpression_MulOverSum(AlgebraicExpression **root) {
-	if(_AlgebraicExpression_IsMultiplicationNode(*root)) {
-		AlgebraicExpression *l = (*root)->operation.children[0];
-		AlgebraicExpression *r = (*root)->operation.children[1];
-
-		// Do not care for (A + B) * (C + D)
-		// As this will end up performing A*C + A*D + B*C + B*D
-		// which is 4 multiplications and 3 additions compared to the original
-		// 2 additions and one multiplication.
-		if((_AlgebraicExpression_IsAdditionNode(l) && !_AlgebraicExpression_IsAdditionNode(r)) ||
-		   (_AlgebraicExpression_IsAdditionNode(r) && !_AlgebraicExpression_IsAdditionNode(l))) {
-
-			AlgebraicExpression *add = AlgebraicExpression_NewOperation(AL_EXP_ADD);
-			AlgebraicExpression *lMul = AlgebraicExpression_NewOperation(AL_EXP_MUL);
-			AlgebraicExpression *rMul = AlgebraicExpression_NewOperation(AL_EXP_MUL);
-
-			AlgebraicExpression_AddChild(add, lMul);
-			AlgebraicExpression_AddChild(add, rMul);
-
-			AlgebraicExpression *A;
-			AlgebraicExpression *B;
-			AlgebraicExpression *C;
-
-			if(l->operation.op == AL_EXP_ADD) {
-				// Lefthand side is addition.
-				// (A + B) * C = (A * C) + (B * C)
-
-				A = _AlgebraicExpression_OperationRemoveLeftmostChild(l);
-				B = _AlgebraicExpression_OperationRemoveRightmostChild(l);
-				C = r;
-
-				AlgebraicExpression_Free(l);
-				AlgebraicExpression_AddChild(lMul, A);
-				AlgebraicExpression_AddChild(lMul, C);
-				AlgebraicExpression_AddChild(rMul, B);
-				AlgebraicExpression_AddChild(rMul, AlgebraicExpression_Clone(C));
-
-				// TODO: Mark r as reusable.
-				// if(r->type == AL_OPERATION) r->operation.reusable = true;
-			} else {
-				// Righthand side is addition.
-				// C * (A + B) = (C * A) + (C * B)
-
-				A = _AlgebraicExpression_OperationRemoveLeftmostChild(r);
-				B = _AlgebraicExpression_OperationRemoveRightmostChild(r);
-				C = l;
-
-				AlgebraicExpression_Free(r);
-				AlgebraicExpression_AddChild(lMul, C);
-				AlgebraicExpression_AddChild(lMul, A);
-				AlgebraicExpression_AddChild(rMul, AlgebraicExpression_Clone(C));
-				AlgebraicExpression_AddChild(rMul, B);
-
-				// TODO: Mark r as reusable.
-				// if(l->type == AL_OPERATION) l->operation.reusable = true;
-			}
-			*root = add;
-			return true;
-		}
-	}
-
-	// Recurse.
-	uint child_count = AlgebraicExpression_ChildCount(*root);
-	for(uint i = 0; i < child_count; i++) {
-		if(__AlgebraicExpression_MulOverSum((*root)->operation.children + i)) return true;
-	}
-	return false;
-}
-
-// Distributive, multiplication over addition:
-// A * (B + C) = (A * B) + (A * C)
-//
-//           (*)
-//   (A)             (+)
-//            (B)          (C)
-//
-// Becomes
-//
-//               (+)
-//       (*)                (*)
-// (A)        (B)     (A)        (C)
-//
-// Whenever we encounter a multiplication operation
-// where one child is an addition operation and the other child
-// is a multiplication operation, we'll replace root multiplication
-// operation with an addition operation with two multiplication operations
-// one for each child of the original addition operation, as can be seen above.
-// we'll want to reuse the left handside of the multiplication.
-static void _AlgebraicExpression_MulOverSum(AlgebraicExpression **root) {
-	// As long as the tree changes keep modifying.
-	while(__AlgebraicExpression_MulOverSum(root));
-}
-
-/* Collect all operands under given root by performing a left to right scan
- * adding each encountered operand to the `operands` array. */
-static void _AlgebraicExpression_CollectOperands(AlgebraicExpression *root,
-												 AlgebraicExpression ***operands) {
-	uint child_count = 0;
-
-	switch(root->type) {
-	case AL_OPERAND:
-		*operands = array_append(*operands, root);
-		break;
-	case AL_OPERATION:
-		switch(root->operation.op) {
-		case AL_EXP_TRANSPOSE:
-			// Transpose is considered as an operand.
-			*operands = array_append(*operands, root);
-			break;
-		case AL_EXP_ADD:
-		case AL_EXP_MUL:
-			child_count = AlgebraicExpression_ChildCount(root);
-			for(uint i = 0; i < child_count; i++) {
-				_AlgebraicExpression_CollectOperands(AlgebraicExpression_Clone(CHILD_AT(root, i)), operands);
-			}
-			break;
-		default:
-			assert("Unknow algebraic expression operation type" && false);
-			break;
-		}
-		break;
-	default:
-		assert("Unknow algebraic expression node type" && false);
-		break;
-	}
-}
-
-/* Collapse multiplication operation under a single multiplication op
- * exp = A * B * C
- * exp can be computed in two ways:
- * 1. (A * B) * C
- * 2. A * (B * C)
- * by flattening the expression both 1 and 2 are represented by the
- * same tree structure. */
-static void _AlgebraicExpression_FlattenMultiplications(AlgebraicExpression *root) {
-	assert(root);
-	uint child_count;
-
-	switch(root->type) {
-	case AL_OPERATION:
-		switch(root->operation.op) {
-		case AL_EXP_ADD:
-		case AL_EXP_TRANSPOSE:
-			// Keep searching for a multiplication operation.
-			child_count = AlgebraicExpression_ChildCount(root);
-			for(int i = 0; i < child_count; i++) {
-				_AlgebraicExpression_FlattenMultiplications(CHILD_AT(root, i));
-			}
-			break;
-
-		case AL_EXP_MUL:
-			// Root has sub multiplication node(s).
-			if(AlgebraicExpression_OperationCount(root, AL_EXP_MUL) > 1) {
-				child_count = AlgebraicExpression_ChildCount(root);
-				AlgebraicExpression **flat_children = array_new(AlgebraicExpression *, child_count);
-				_AlgebraicExpression_CollectOperands(root, &flat_children);
-				// Free `old` child array.
-				for(uint i = 0; i < child_count; i++) AlgebraicExpression_Free(CHILD_AT(root, i));
-				array_free(root->operation.children);
-				root->operation.children = flat_children;
-			}
-
-			break;
-		default:
-			assert("Unknown algebraic operation type" && false);
-			break;
-		}
-	default:
-		break;
-	}
-}
-
-/* Multiplies `exp` to the left by `lhs`.
- * Returns new expression root.
- * `lhs` = (A + B)
- * `exp` = Transpose(C)
- * Returns (A + B) * Transpose(C) where `*` is the new root. */
-static AlgebraicExpression *_AlgebraicExpression_MultiplyToTheLeft
-(
-	AlgebraicExpression *lhs,
-	AlgebraicExpression *exp
-) {
-	assert(lhs && exp);
-	AlgebraicExpression *mul = AlgebraicExpression_NewOperation(AL_EXP_MUL);
-	AlgebraicExpression_AddChild(mul, lhs);
-	AlgebraicExpression_AddChild(mul, exp);
-	return mul;
-}
-
-/* Multiplies `exp` to the right by `rhs`.
- * Returns new expression root.
- * `exp` = Transpose(C)
- * `rhs` = (A + B)
- * Returns Transpose(C) * (A + B) where `*` is the new root. */
-static AlgebraicExpression *_AlgebraicExpression_MultiplyToTheRight
-(
-	AlgebraicExpression *exp,
-	AlgebraicExpression *rhs
-) {
-	assert(exp && rhs);
-	AlgebraicExpression *mul = AlgebraicExpression_NewOperation(AL_EXP_MUL);
-	AlgebraicExpression_AddChild(mul, exp);
-	AlgebraicExpression_AddChild(mul, rhs);
-	return mul;
 }
 
 /* Adds `exp` to the left by `lhs`.
@@ -318,422 +70,6 @@ static AlgebraicExpression *_AlgebraicExpression_AddToTheRight
 	return add;
 }
 
-//------------------------------------------------------------------------------
-// Static internal functions.
-//------------------------------------------------------------------------------
-
-/* Node with (income + outcome degree) > 2
- * is considered a highly connected node. */
-static bool _highly_connected_node(const QueryGraph *qg, const char *alias) {
-	// Look up node in qg.
-	QGNode *n = QueryGraph_GetNodeByAlias(qg, alias);
-	return ((array_len(n->incoming_edges) + array_len(n->outgoing_edges)) > 2);
-}
-
-static inline bool _referred_entity(const char *alias) {
-	AST *ast = QueryCtx_GetAST();
-	return AST_AliasIsReferenced(ast, alias);
-}
-
-/* If the edge is referenced or of a variable length, it should populate the AlgebraicExpression. */
-static inline bool _should_populate_edge(QGEdge *e) {
-	return (_referred_entity(e->alias) || QGEdge_VariableLength(e));
-}
-
-/* Checks if given expression contains a variable length edge. */
-static bool _AlgebraicExpression_ContainsVariableLengthEdge
-(
-	const QueryGraph *qg,
-	const AlgebraicExpression *root
-) {
-	uint child_count = 0;
-	switch(root->type) {
-	case AL_OPERATION:
-		child_count = AlgebraicExpression_ChildCount(root);
-		for(uint i = 0; i < child_count; i++) {
-			if(_AlgebraicExpression_ContainsVariableLengthEdge(qg, root->operation.children[i])) return true;
-		}
-		return false;
-	case AL_OPERAND:
-		if(root->operand.edge) {
-			QGEdge *e = QueryGraph_GetEdgeByAlias(qg, root->operand.edge);
-			return QGEdge_VariableLength(e);
-		}
-		break;
-	default:
-		assert("Unknow algebraic expression node type" && false);
-	}
-	return false;
-}
-
-static void _RemovePathFromGraph(QueryGraph *g, QGEdge **path) {
-	uint edge_count = array_len(path);
-	for(uint i = 0; i < edge_count; i++) {
-		QGEdge *e = path[i];
-		QGNode *src = e->src;
-		QGNode *dest = e->dest;
-
-		QueryGraph_RemoveEdge(g, e);
-		QGEdge_Free(e);
-		if(QGNode_EdgeCount(src) == 0) {
-			QueryGraph_RemoveNode(g, src);
-			QGNode_Free(src);
-		}
-		if(src != dest && QGNode_EdgeCount(dest) == 0) {
-			QueryGraph_RemoveNode(g, dest);
-			QGNode_Free(dest);
-		}
-	}
-}
-
-/* Variable length expression must contain only a single operand, the edge being
- * traversed multiple times, in cases such as (:labelA)-[e*]->(:labelB) both label A and B
- * are applied via a label matrix operand, this function migrates A and B from a
- * variable length expression to other expressions. */
-static AlgebraicExpression **_AlgebraicExpression_IsolateVariableLenExps(
-	const QueryGraph *qg, AlgebraicExpression **expressions) {
-	/* Return value is a new set of expressions, where each variable length expression
-	 * is guaranteed to have a single operand, as such in the worst case the number of
-	 * expressions doubles + 1. */
-	size_t expCount = array_len(expressions);
-	AlgebraicExpression **res = array_new(AlgebraicExpression *, expCount * 2 + 1);
-
-	/* Scan through each expression, locate expression which
-	 * have a variable length edge in them. */
-	for(size_t expIdx = 0; expIdx < expCount; expIdx++) {
-		AlgebraicExpression *exp = expressions[expIdx];
-		if(!_AlgebraicExpression_ContainsVariableLengthEdge(qg, exp)) {
-			res = array_append(res, exp);
-			continue;
-		}
-
-		// Expression contains a variable length edge.
-		QGNode *src = QueryGraph_GetNodeByAlias(qg, AlgebraicExpression_Source(exp));
-		QGNode *dest = QueryGraph_GetNodeByAlias(qg, AlgebraicExpression_Destination(exp));
-
-		// A variable length expression with a labeled source node
-		// We only care about the source label matrix, when it comes to
-		// the first expression, as in the following expressions
-		// src is the destination of the previous expression.
-		if(expIdx == 0 && src->label) {
-			// Remove src node matrix from expression.
-			AlgebraicExpression *op = AlgebraicExpression_RemoveLeftmostNode(exp);
-			res = array_append(res, op);
-		}
-
-		res = array_append(res, exp);
-
-		// If the expression has a labeled destination, separate it into its own expression.
-		if(dest->label) {
-			// Remove dest node matrix from expression.
-			AlgebraicExpression *op = AlgebraicExpression_RemoveRightmostNode(exp);
-
-			/* See if dest mat can be prepended to the following expression.
-			 * If not create a new expression. */
-			if(expIdx < expCount - 1 &&
-			   !_AlgebraicExpression_ContainsVariableLengthEdge(qg, expressions[expIdx + 1])) {
-				_AlgebraicExpression_MultiplyToTheLeft(expressions[expIdx + 1], op);
-			} else {
-				res = array_append(res, op);
-			}
-		}
-	}
-
-	array_free(expressions);
-	return res;
-}
-
-static inline void _AlgebraicExpression_Execute_MUL(GrB_Matrix C, GrB_Matrix A, GrB_Matrix B,
-													GrB_Descriptor desc) {
-	// A,B,C must be boolean matrices.
-	GrB_Info res = GrB_mxm(
-					   C,                   // Output
-					   GrB_NULL,            // Mask
-					   GrB_NULL,            // Accumulator
-					   GxB_LOR_LAND_BOOL,   // Semiring
-					   A,                   // First matrix
-					   B,                   // Second matrix
-					   desc                 // Descriptor
-				   );
-
-	if(res != GrB_SUCCESS) {
-		// If the multiplication failed, print error info to stderr and exit.
-		fprintf(stderr, "Enountered error in matrix multiplication:\n%s\n", GrB_error());
-		assert(false);
-	}
-}
-
-static inline bool _should_divide_expression(QGEdge **path, int idx, const QueryGraph *qg) {
-	QGEdge *e = path[idx];
-
-	return (_should_populate_edge(e)                    ||  // This edge is populated.
-			_should_populate_edge(path[idx + 1])        ||  // The next edge is populated.
-			_highly_connected_node(qg, e->dest->alias)  ||  // Destination node in+out degree > 2.
-			_referred_entity(e->dest->alias));              // Destination node is referenced.
-}
-
-/* Break down path into sub paths.
- * Considering referenced intermidate nodes and edges. */
-static QGEdge ***_Intermediate_Paths
-(
-	QGEdge **path,
-	const QueryGraph *qg
-) {
-	QGEdge *e = NULL;
-	int pathLen = array_len(path);
-
-	/* Allocating maximum number of expression possible. */
-	QGEdge ***paths = array_new(QGEdge **, pathLen);
-	QGEdge **intermediate_path = array_new(QGEdge *, 1);
-
-	/* Scan path left to right,
-	 * construct intermidate paths by "breaking" on referenced entities. */
-	for(int i = 0; i < pathLen - 1; i++) {
-		e = path[i];
-		intermediate_path = array_append(intermediate_path, e);
-		if(_should_divide_expression(path, i, qg)) {
-			// Break! add current path to paths and create a new path.
-			paths = array_append(paths, intermediate_path);
-			intermediate_path = array_new(QGEdge *, 1);
-		}
-	}
-
-	// Handle last hop.
-	e = path[pathLen - 1];
-	intermediate_path = array_append(intermediate_path, e);
-	// If `intermidate_path` contains just the last hop it's not part of paths, add it.
-	if(array_len(intermediate_path) == 1) paths = array_append(paths, intermediate_path);
-
-	return paths;
-}
-
-static AlgebraicExpression *_AlgebraicExpression_OperandFromNode
-(
-	QGNode *n
-) {
-	GrB_Matrix mat;
-	bool free = false;
-	bool diagonal = true;
-	bool transpose = false;
-	Graph *g = QueryCtx_GetGraph();
-
-	if(n->labelID == GRAPH_UNKNOWN_LABEL) mat = Graph_GetZeroMatrix(g);
-	else mat = Graph_GetLabelMatrix(g, n->labelID);
-
-	return AlgebraicExpression_NewOperand(mat, free, diagonal, n->alias, n->alias, NULL);
-}
-
-static AlgebraicExpression *_AlgebraicExpression_OperandFromEdge
-(
-	QGEdge *e,
-	bool transpose
-) {
-	GrB_Matrix mat;
-	uint reltype_id;
-	Graph *g = QueryCtx_GetGraph();
-	AlgebraicExpression *add;
-	AlgebraicExpression *root = NULL;
-
-	uint reltype_count = array_len(e->reltypeIDs);
-	switch(reltype_count) {
-	case 0: // No relationship types specified; use the full adjacency matrix
-		mat = Graph_GetAdjacencyMatrix(g);
-		root = AlgebraicExpression_NewOperand(mat, false, false, e->src->alias, e->dest->alias,
-											  e->alias);
-		break;
-	case 1: // One relationship type
-		reltype_id = e->reltypeIDs[0];
-		if(reltype_id == GRAPH_UNKNOWN_RELATION) mat = Graph_GetZeroMatrix(g);
-		else mat = Graph_GetRelationMatrix(g, e->reltypeIDs[0]);
-		root = AlgebraicExpression_NewOperand(mat, false, false, e->src->alias, e->dest->alias,
-											  e->alias);
-		break;
-	default: // Multiple edge type: -[:A|:B]->
-		add = AlgebraicExpression_NewOperation(AL_EXP_ADD);
-		for(uint i = 0; i < reltype_count; i++) {
-			uint reltype_id = e->reltypeIDs[i];
-			// No matrix to add
-			if(reltype_id == GRAPH_UNKNOWN_RELATION) mat = Graph_GetZeroMatrix(g);
-			else mat = Graph_GetRelationMatrix(g, reltype_id);
-			AlgebraicExpression *operand = AlgebraicExpression_NewOperand(mat, false, false,
-																		  e->src->alias, e->dest->alias, e->alias);
-			AlgebraicExpression_AddChild(add, operand);
-		}
-		root = add;
-		break;
-	}
-
-	/* Expand fixed variable length edge.
-	 * -[A*2..2]->
-	 * A*A
-	 * -[A|B*2..2]->
-	 * (A+B) * (A+B) */
-	if(!QGEdge_VariableLength(e) && e->minHops > 1) {
-		AlgebraicExpression *mul = AlgebraicExpression_NewOperation(AL_EXP_MUL);
-		AlgebraicExpression_AddChild(mul, root);
-		for(int i = 1; i < e->minHops; i++) {
-			// Clone to avoid double free.
-			AlgebraicExpression_AddChild(mul, AlgebraicExpression_Clone(root));
-		}
-		root = mul;
-	}
-
-	if(transpose) {
-		AlgebraicExpression *op_transpose = AlgebraicExpression_NewOperation(AL_EXP_TRANSPOSE);
-		AlgebraicExpression_AddChild(op_transpose, root);
-		root = op_transpose;
-	}
-
-	// If src node has a label, multiply to the left by label matrix.
-	if(e->src->label) {
-		root = _AlgebraicExpression_MultiplyToTheLeft(_AlgebraicExpression_OperandFromNode(e->src), root);
-	}
-
-	return root;
-}
-
-/* In case edges `a` and `b` share a node:
- * (a)-[E0]->(b)<-[E1]-(c)
- * than the shared entity is returned
- * if edges are disjoint, NULL is returned. */
-static QGNode *_SharedNode
-(
-	const QGEdge *a,
-	const QGEdge *b
-) {
-	assert(a && b);
-	if(a->dest == b->src) return a->dest;   // (a)-[E0]->(b)-[E1]->(c)
-	if(a->src == b->dest) return a->src;    // (a)<-[E0]-(b)<-[E1]-(c)
-	if(a->src == b->src) return a->src;     // (a)<-[E0]-(b)-[E1]->(c)
-	if(a->dest == b->dest) return a->dest;  // (a)-[E0]->(b)<-[E1]-(c)
-	return NULL;
-}
-
-static void _reversePath
-(
-	QGEdge **path,
-	uint path_len,
-	bool *transpositions
-) {
-	for(uint i = 0; i < path_len; i++) {
-		/* A reversed path should have its transpositions flipped
-		 * transpose(transpose(A)) = A */
-		transpositions[i] = !transpositions[i];
-	}
-
-	/* Transpose(A*B) = Transpose(B) * Transpose(A)
-	 * (a)<-[A]-(b)<-[B]-(c)-[C]->(d)
-	 * At * Bt * C
-	 * Transpose(At * Bt * C) =
-	 * = Ct * B * A
-	 * (d)-[Ct]->(c)-[B]->(b)-[A]->(a) */
-
-	// Reverse the path array as well as the transposition array
-	for(uint i = 0; i < path_len / 2; i++) {
-		uint opposite = path_len - i - 1;
-		QGEdge *tmp = path[opposite];
-		path[opposite] = path[i];
-		path[i] = tmp;
-
-		bool transpose_tmp = transpositions[opposite];
-		transpositions[opposite] = transpositions[i];
-		transpositions[i] = transpose_tmp;
-	}
-}
-
-static void _normalizePath
-(
-	QGEdge **path,          // Path to normalize.
-	uint path_len,          // Path length.
-	bool *transpositions    // Specifies which edges need to be transposed.
-) {
-	// Initialize `transpositions` array.
-	for(uint i = 0; i < path_len; i++) transpositions[i] = false;
-
-	// A single leg path.
-	if(path_len <= 1) return;
-
-	uint transposeCount = 0;
-	// For every edge except the last:
-	for(uint i = 0; i < path_len - 1; i++) {
-		QGEdge *e = path[i];
-		QGEdge *follow = path[i + 1];
-		QGNode *shared = _SharedNode(e, follow);
-		assert(shared);
-
-		/* The edge should be transposed if its destination is not shared.
-		 * (dest)<-[e]-(shared)-[follow]->()
-		 * (dest)<-[e]-(shared)<-[follow]-() */
-		if(e->dest != shared) {
-			transpositions[i] = true;
-			transposeCount++;
-		}
-	}
-
-	// For the last edge, transpose if its source is not shared.
-	QGEdge *e = path[path_len - 1];
-	QGNode *shared = _SharedNode(path[path_len - 2], e);
-	if(e->src != shared) {
-		transposeCount++;
-		transpositions[path_len - 1] = true;
-	}
-
-	// Reverse entire path if the majority of edges must be transposed.
-	if(transposeCount > (path_len - transposeCount)) {
-		_reversePath(path, path_len, transpositions);
-	}
-
-	// Apply transpose.
-	for(uint i = 0; i < path_len; i++) {
-		QGEdge *e = path[i];
-		if(transpositions[i]) QGEdge_Reverse(e);
-	}
-}
-
-static AlgebraicExpression *_AlgebraicExpression_FromPath
-(
-	QGEdge **path
-) {
-	assert(path);
-
-	QGEdge *e = NULL;
-	uint path_len = array_len(path);
-	assert(path_len > 0);
-	AlgebraicExpression *root = NULL;
-
-	/* Treating path as a chain
-	 * we're aligning all edges to "point right"
-	 * (A)-[E0]->(B)-[E0]->(C)-[E0]->(D).
-	 * e.g.
-	 * (A)-[E0]->(B)<-[E1]-(C)-[E2]->(D)
-	 * E1 will be transposed:
-	 * (A)-[E0]->(B)-[E1]->(C)-[E2]->(D) */
-
-	bool transpositions[path_len];
-	_normalizePath(path, path_len, transpositions);
-
-	// Construct expression.
-	for(int i = 0; i < path_len; i++) {
-		e = path[i];
-		// Add Edge matrix.
-		AlgebraicExpression *op = _AlgebraicExpression_OperandFromEdge(e, transpositions[i]);
-
-		if(!root) {
-			root = op;
-		} else {
-			// Connect via a multiplication node.
-			root = _AlgebraicExpression_MultiplyToTheRight(root, op);
-		}
-	}   // End of path traversal.
-
-	// If last node on path has a label, multiply by label matrix.
-	if(e->dest->label) {
-		root = _AlgebraicExpression_MultiplyToTheRight(root, _AlgebraicExpression_OperandFromNode(e->dest));
-	}
-
-	return root;
-}
-
 static AlgebraicExpression *_AlgebraicExpression_CloneOperation
 (
 	const AlgebraicExpression *exp
@@ -754,127 +90,6 @@ static AlgebraicExpression *_AlgebraicExpression_CloneOperand
 	return AlgebraicExpression_NewOperand(exp->operand.matrix, exp->operand.free,
 										  exp->operand.diagonal, exp->operand.src,
 										  exp->operand.dest, exp->operand.edge);
-}
-
-//------------------------------------------------------------------------------
-// AlgebraicExpression construction.
-//------------------------------------------------------------------------------
-
-// Construct algebraic expression form query graph.
-AlgebraicExpression **AlgebraicExpression_FromQueryGraph
-(
-	const QueryGraph *qg,   // Query-graph to process
-	uint *exp_count         // Number of algebraic expressions generated.
-) {
-	assert(qg && exp_count);
-
-	/* Construct algebraic expression(s) from query-graph.
-	 * Trying to take advantage of long multiplications with as few
-	 * transpose as possible we'll transform paths crossing the graph
-	 * "diameter", these are guarantee to be the longest, although
-	 * there might be situations in which these are not the most optimal paths
-	 * to explore.
-	 *
-	 * Once a path been transformed it's removed from the query-graph and the process
-	 * repeat itself. */
-
-	/* A graph with no edges implies an empty algebraic expression
-	 * the reasoning behind this decission is that algebraic expression
-	 * represent graph traversals, no edges means no traversals. */
-	uint edge_count = QueryGraph_EdgeCount(qg);
-	if(edge_count == 0) {
-		*exp_count = 0;
-		return NULL;
-	}
-
-	bool acyclic = IsAcyclicGraph(qg);
-	QueryGraph *g = QueryGraph_Clone(qg);
-	AlgebraicExpression **exps = array_new(AlgebraicExpression *, 1);
-
-	// As long as the query-graph isn't empty.
-	while(QueryGraph_EdgeCount(g) > 0) {
-		// Get leaf nodes at the deepest level.
-		int depth;
-		QGNode *n;
-		if(acyclic) n = LongestPathTree(g, &depth); // Graph is a tree.
-		else n = LongestPathGraph(g, &depth);       // Graph contains cycles.
-
-		// Get a path of length level, allow closing a cycle if the graph is not acyclic.
-		QGEdge **path = DFS(n, depth, !acyclic);
-		uint path_len = array_len(path);
-		assert(path_len == depth);
-
-		/* TODO:
-		 * In case path is a cycle, e.g. (b)-[]->(a)-[]->(b)
-		 * make sure the first node on the path is referenced, _should_divide_expression(path, 0) is true.
-		 * if this is not the case we will unnecessarily break the generated expression into 2 sub expressions
-		 * while what we can do is simply rotate the cycle, (a)-[]->(b)-[]->(a)
-		 * this is exactly the same only now we won't sub divide.
-		 * Checking if path is a cycle done by testing the start and end node. */
-
-		// Split path into sub paths.
-		QGEdge ***paths = _Intermediate_Paths(path, qg);
-		AlgebraicExpression **sub_exps = array_new(AlgebraicExpression *, 1);
-
-		uint path_count = array_len(paths);
-		for(uint i = 0; i < path_count; i++) {
-			// Construct expression.
-			AlgebraicExpression *exp = _AlgebraicExpression_FromPath(paths[i]);
-			sub_exps = array_append(sub_exps, exp);
-
-			/* Remove exp[i] src label matrix (left most operand) as it's
-			 * being used by exp[i-1] dest label matrix.
-			 * (:A)-[:X]->(:B)-[:Y]->(:C)
-			 * exp0: A * X * B
-			 * exp1: B * Y * C
-			 * should become
-			 * exp0: A * X * B
-			 * exp1: Y * C */
-			if(i > 0) {
-				QGNode *src = QueryGraph_GetNodeByAlias(qg, AlgebraicExpression_Source(exp));
-				if(src->label) {
-					/* exp[i] shares a label matrix with exp[i-1]
-					 * remove redundancy. */
-					AlgebraicExpression *redundent = AlgebraicExpression_RemoveLeftmostNode(exp);
-					AlgebraicExpression_Free(redundent);
-				}
-			}
-			// Expression can not be empty.
-			assert(AlgebraicExpression_ChildCount(exp) > 0);
-		}
-
-		sub_exps = _AlgebraicExpression_IsolateVariableLenExps(qg, sub_exps);
-
-		uint sub_count = array_len(sub_exps);
-		for(uint i = 0; i < sub_count; i++) {
-			AlgebraicExpression *exp = sub_exps[i];
-			AlgebraicExpression_Optimize(&exp);
-			// Add constructed expression to return value.
-			exps = array_append(exps, exp);
-		}
-
-		// Remove path from graph.
-		_RemovePathFromGraph(g, path);
-
-		// Clean up
-		for(uint i = 0; i < path_count; i++) array_free(paths[i]);
-		array_free(paths);
-		array_free(sub_exps);
-
-		/* If original graph contained a cycle
-		 * see now after we've removed a path if this is still the case. */
-		if(!acyclic) acyclic = IsAcyclicGraph(g);
-	}
-
-	// TODO just return exps?
-	*exp_count = array_len(exps);
-	AlgebraicExpression **res = rm_malloc(sizeof(AlgebraicExpression *) * (*exp_count));
-	for(size_t i = 0; i < (*exp_count); i++) res[i] = exps[i];
-	array_free(exps);
-
-	QueryGraph_Free(g);
-
-	return res;
 }
 
 //------------------------------------------------------------------------------
@@ -929,6 +144,7 @@ AlgebraicExpression *AlgebraicExpression_Clone
 		break;
 	default:
 		assert("Unknow algebraic expression node type" && false);
+		break;
 	}
 	return NULL;
 }
@@ -1018,6 +234,7 @@ uint AlgebraicExpression_OperandCount
 		break;
 	default:
 		assert("Unknow algebraic expression node type" && false);
+		break;
 	}
 	return operand_count;
 }
@@ -1241,18 +458,6 @@ bool AlgebraicExpression_DiagonalOperand
 	const AlgebraicExpression *op = _AlgebraicExpression_GetOperand(root, operand_idx);
 	assert(op && op->type == AL_OPERAND);
 	return op->operand.diagonal;
-}
-
-//------------------------------------------------------------------------------
-// AlgebraicExpression optimizations
-//------------------------------------------------------------------------------
-void AlgebraicExpression_Optimize
-(
-	AlgebraicExpression **exp
-) {
-	assert(exp);
-	_AlgebraicExpression_MulOverSum(exp);
-	_AlgebraicExpression_FlattenMultiplications(*exp);
 }
 
 //------------------------------------------------------------------------------
